Fix: 로그 데이터 Integer로 수정, 데이터 없어도 0으로 나오게 수정

See merge request s11-bigdata-dist-sub1/S11P21B206!129

+++ backend/kid-service/src/main/java/com/helloword/kidservice/domain/kid/dto/response/LearningStatsResponseDto.java
@@ -11,6 +11,6 @@ public class LearningStatsResponseDto {
     private Integer todayPlayTime;
     private Integer todayCompletedGames;
     private Map<String, Integer> dailyCorrectWordCounts;
-    private Map<String, Double> globalDailyAverageCorrectWordCounts;
+    private Map<String, Integer> globalDailyAverageCorrectWordCounts;
 }
